feat-hotfix: routes analyze, robot, route and point

+++ src/backend/point/model.py
@@ -1,34 +1,29 @@
-import bcrypt
 from datetime import datetime
-
+from __init__ import db
 from prisma import Prisma
 
-def connect_db() -> Prisma:
-        # from app import db as prisma_config
-        db = Prisma()
-        db.connect()
-        return db
-
-class Point_models:
-    def create_point(pointX: float, pointY: float, routeId: int) -> bool:
-        db = connect_db()
-        
-        data = {
+def create_point(pointX: float, pointY: float, routeId: int) -> bool:   
+    data = {
                 'pointX': pointX,
                 'pointY': pointY,
                 'routeId': routeId,
                 'created_at': datetime.now()
-                }
-        db.point.create(data=data)
-        return True
+            }
+    db.point.create(data=data)
+    return True
     
-    def get_points(routeId: int) -> list:
-        db = connect_db()
-        points = db.point.find_many(where={'routeId': routeId})
+def get_points(routeId: int) -> list[Prisma.robot]:
+    points = db.point.find_many(where={'routeId': routeId})
+    if points == None:
+        return []
+    if points is not None:
         return points
     
-    def delete_points(routeId: int) -> bool:
-        db = connect_db()
+def delete_points(routeId: int) -> bool:
+    point = db.point.find_many(where={'routeId': routeId})
+    if not point:
+        return False
+    else:
         db.point.delete_many(where={'routeId': routeId})
         return True
